tree: track tainted configurations
This is in preparation for the Commit() implementation: We don't need
to write files, if their content has not changed.

Implementation detail: only tree methods should be allowed to change
the tainted flag. The config methods could change that flag, but then
all section and option should to that as well for consitency. However,
we don't have back references to the parent object (option→section,
or section→config), so this isn't possible. Also: the parser makes
extensive use of the config/section/option methods, but at that point,
we don't need to keep track of the tainted status (as we're just
building the config instance).