uci: test Commit
This first replaces plain os.* and ioutil.* calls with
a type combining all methods needed. In the test,
that type is then mocked (and spied upon).

Testing file system ops is a bit tricky in Go... I'd usually
use spf13/afero in such cases, but that might increase
the binary size a bit too much (I did not measure this
though).