fix empty option parsing [WIP]
UCI quietly accepts the following config files:

    $ cat /etc/config/test
    config section
    	option empty1
    	option empty2 ''
    	list empty3
    	list empty4 ''

    # uci export test
    package test

    config section
    	list empty3 ''
    	list empty4 ''

whereas go-uci returns a parse error. Note how test.section.empty1
and .empty2 are completely ignored with UCI.

This commit is a first attempt on matching the behaviour. The lexer
is already capable of capturing empty values, but the parser doesn't
handle this case (yet).

ref #13