fix: set/generate correct option type
An UCI option has one of two types: it is either a single-value
option, or it is a multi-value list (lists with one entry are
valid as well).

The previous API did not do the right thing when writing options
back to file: it looked only at the number of entries for an
option and may have inadvertently "converted" lists to options.
When reading back a UCI config, some services complained that
they expected a "list", but got an "option" instead.

This commit introduces a new API method to explicitly define
the type when setting options:

    uci.SetType("cfg", "sect", "opt", uci.TypeOption, "v1")
    uci.SetType("cfg", "sect", "opt", uci.TypeList, "v1", "v2", "v3")

It should be mentioned, that

    uci.SetType("cfg", "sect", "opt", uci.TypeOption, "v1", "v2", "v3")

will not panic(), but the generated UCI config may be invalid
nonetheless.

This commit hereby deprecates the `uci.Set()`. It is left for backward
compatability.

Fixes #16